pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sol-trader"
version = "0.1.0"
description = "Solana meme coin paper trading bot with Telegram signal ingestion"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.36.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["trader*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
